+++ moqt-core/src/modules/handlers.rs
@@ -1,6 +1,6 @@
 pub(crate) mod announce_handler;
 pub(crate) mod server_setup_handler;
-pub(crate) mod subscribe_request_handler;
-pub(crate) mod subscribe_response_handler;
+pub(crate) mod subscribe_handler;
+pub(crate) mod subscribe_ok_handler;
 pub(crate) mod unannounce_handler;
 pub(crate) mod unsubscribe_handler;
